test-video-sharing/backend/requirements.txt
anyio==4.13.0
bcrypt==5.0.0
blinker==1.9.0
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.3
cryptography==47.0.0
Flask==3.1.3
flask-cors==6.0.2
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.13
invoke==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.1.0
MarkupSafe==3.0.3
paramiko==4.0.0
pycparser==3.0
PyNaCl==1.6.2
python-dotenv==1.2.2
python-gvm==27.0.2
requests==2.33.1
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.8
